refactor(tests): extract test module from main.rs into separate files
- Create src/lib.rs exposing core API (apply_overlay, remove_overlay, show_status)
- Move ~1,600 lines of inline tests to tests/ directory
- Add TestContext builder in tests/common/ to reduce test setup duplication
- Split tests into unit/, integration/, and cli.rs
- Slim main.rs from ~2,000 lines to ~535 lines (CLI-only)
