+++ tests/integration/mod.rs
@@ -1,0 +1,3 @@
+mod apply;
+mod remove;
+mod status;
