+++ tests/tests.rs
@@ -1,0 +1,5 @@
+//! Integration test suite for repoverlay.
+
+mod common;
+mod integration;
+mod unit;
