+++ tests/unit/mod.rs
@@ -1,0 +1,1 @@
+mod remove_section;
